requirements.txt
streamlit
fastapi
uvicorn
sqlalchemy
pandas
plotly
requests
python-dotenv
openai
google-generativeai
pypdf
ezdxf
matplotlib
opencv-python
ultralytics
reportlab
arabic-reshaper
python-bidi
python-jose[cryptography]
passlib[bcrypt]
exifread
geopy
redis
apscheduler
mistralai
pytest
pytest-asyncio
azure-identity
azure-keyvault-secrets